+++ code/src/timer.h
@@ -1,4 +1,4 @@
 #pragma once
 
-void timer_enable(void (*timer_callback)(void), unsigned timer_freq);
+void timer_enable(void (*timer_callback)(void), float timer_freq);
 void timer_disable(void);
